requirements.txt
Flask
Flask-Session
cs50
python-dotenv
Flask-Bcrypt
Authlib